Compressible version still blows up after finite time with systematic oscillations. After correction of coefficients in the trapezoidal rule scheme in the explicit predictor, the psinc version still runs, provided there is a background wind, but with worse symmetry behaviour in the left and right running pulses.   I have now checked the vertical velocity fluctuations over several cycles of the vertical predictor alone in implicit trapezoidal mode, and it does seem to work rather well.   My hunch is that this asymmetry is due to my embedding the vertical predictor step in the middle of the directional opsplit. It is then clear that the pressure and buoyancy fields get displaced relative to each other by the background advection both in the predictor and in the corrector. This may induce the observed asymmetry. Therefore, I will next run the vertical predictor first and last in the opsplit cycle ... .   To this end, it becomes necessary to split the explicit part of the implicit trapezoidal rule into - first Euler-backward, then Euler-forward. I have realized that option in the gravity discretisation of the vertical predictor now and roughly tested it with both vertical steps over delta t/2 strides still in the middle. It produces the same results as with one full time step vertical predictor step that makes a full delta t stride.

+++ RKLM_Reference/Physics/Gasdynamics/recovery.h
@@ -33,7 +33,7 @@ void recovery_gravity(
 					  const double lambda, 
 					  const int nmax,
                       const int stage,
-                      const int implicit);
+                      const double implicit);
 
 /*------------------------------------------------------------------------------
